fastapi/requirements.txt
# Assuming python==3.12
fastapi[standard]
azure-identity
psycopg2-binary
sqlalchemy
langsmith==0.1.139
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-openai==0.2.3
langgraph==0.2.44
langchain-community==0.3.3
langchain-google-genai==2.0.0
langchain_milvus
boto3
openai>=1.54.0
tenacity==8.2.3
requests 
pytesseract 
pillow
langchain-openai
python-dotenv
requests
beautifulsoup4
chardet
pymilvus==2.5.0
unidecode
python-docx
mammoth
openpyxl
pymupdf
tiktoken
markdown2